requirements.txt
fastapi
uvicorn
python-multipart
requests
qdrant-client
sentence-transformers
pypdf
python-docx
pandas
numpy

